requirements.txt
absl-py==0.6.1
astor==0.7.1
backports.functools-lru-cache==1.5
backports.weakref==1.0.post1
cycler==0.10.0
enum34==1.1.6
funcsigs==1.0.2
futures==3.2.0
gast==0.2.0
grpcio==1.17.1
h5py==2.9.0
Keras==2.2.4
Keras-Applications==1.0.6
Keras-Preprocessing==1.0.5
kiwisolver==1.0.1
Markdown==3.0.1
matplotlib==2.2.3
mock==2.0.0
numpy==1.15.4
pbr==5.1.1
Pillow==5.3.0
protobuf==3.6.1
pycairo==1.18.0
pyparsing==2.3.0
python-dateutil==2.7.5
pytz==2018.7
PyYAML==3.13
scipy==1.2.0
six==1.12.0
subprocess32==3.5.3
tensorboard==1.12.1
tensorflow==1.12.0
termcolor==1.1.0
Werkzeug==0.14.1
